JOE-884: Split Sidebar, cowork-api, and event-message-handlers mega-files
Extract Sidebar branding and workspace switcher into dedicated components,
split browser cowork-api into transport/SSE hub and channels modules, and
move session.next native message handlers into their own desktop module.
Ratchet renderer line budgets under the general 900-line limit.
